Move fake event reporter to its own file
